pyproject.toml
[tool.poetry]
name = "codeinterpreterapi"
version = "0.0.9"
description = "CodeInterpreterAPI is an (unofficial) open source python interface for the ChatGPT CodeInterpreter."
authors = ["Shroominic <contact@shroominic.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
langchain = "^0.0.242"
codeboxapi = "^0.0.14"

[tool.poetry.extras]
all = ["jupyter-kernel-gateway", "streamlit", "Pillow"]
localbox = ["jupyter-kernel-gateway"]
frontend = ["streamlit"]
image_support = ["Pillow"]

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
mypy = "^1.4.1"
black = "^23.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
